// recovery-office/src/types/api.types.ts
export enum ServiceType {
  INITIAL_CONSULTATION = 'initial_consultation',
  FOLLOW_UP = 'follow_up',
  COMPREHENSIVE_ASSESSMENT = 'comprehensive_assessment',
  SPECIALIZED_TREATMENT = 'specialized_treatment',
  EMERGENCY_SESSION = 'emergency_session',
}
// ...
export interface TimeSlot {
  startTime: string;
  endTime: string;
  isAvailable: boolean;
}